requirements.txt
Django==3.1.7
PyMySQL==1.0.2
openai==0.28.0
PyPDF2==3.0.1
sqlparse==0.5.1
pillow==11.0.0
pytesseract==0.3.13
python-docx==1.1.2
python-dateutil==2.9.0.post0
django-storages
google-cloud-storage
djangorestframework==3.15.2